MySQL release 5.7.24 26

+++ client/mysqltest.cc
@@ -1,4 +1,4 @@
-/* Copyright (c) 2000, 2017, Oracle and/or its affiliates. All rights reserved.
+/* Copyright (c) 2000, 2018, Oracle and/or its affiliates. All rights reserved.
 
    This program is free software; you can redistribute it and/or modify
    it under the terms of the GNU General Public License as published by
@@ -50,6 +50,7 @@
 
 #include <welcome_copyright_notice.h> // ORACLE_WELCOME_COPYRIGHT_NOTICE
 
+#include <sstream>
 #include <string>
 #include <algorithm>
 #include <functional>
@@ -164,9 +165,6 @@ static my_bool is_windows= 0;
 static char **default_argv;
 static const char *load_default_groups[]= { "mysqltest", "client", 0 };
 static char line_buffer[MAX_DELIMITER_LENGTH], *line_buffer_pos= line_buffer;
-#if !defined(HAVE_YASSL)
-static const char *opt_server_public_key= 0;
-#endif
 static my_bool can_handle_expired_passwords= TRUE;
 
 /* Info on properties that can be set with --enable_X and --disable_X */
@@ -298,6 +296,7 @@ typedef Prealloced_array<st_command*, 1024> Q_lines;
 Q_lines *q_lines;
 
 #include "sslopt-vars.h"
+#include <caching_sha2_passwordopt-vars.h>
 
 struct Parser
 {
@@ -388,7 +387,7 @@ enum enum_commands {
   Q_ENABLE_INFO, Q_DISABLE_INFO,
   Q_ENABLE_SESSION_TRACK_INFO, Q_DISABLE_SESSION_TRACK_INFO,
   Q_ENABLE_METADATA, Q_DISABLE_METADATA,
-  Q_EXEC, Q_EXECW, Q_DELIMITER,
+  Q_EXEC, Q_EXECW, Q_EXEC_BACKGROUND, Q_DELIMITER,
   Q_DISABLE_ABORT_ON_ERROR, Q_ENABLE_ABORT_ON_ERROR,
   Q_DISPLAY_VERTICAL_RESULTS, Q_DISPLAY_HORIZONTAL_RESULTS,
   Q_QUERY_VERTICAL, Q_QUERY_HORIZONTAL, Q_SORTED_RESULT,
@@ -402,6 +401,7 @@ enum enum_commands {
   Q_WRITE_FILE, Q_COPY_FILE, Q_PERL, Q_DIE, Q_EXIT, Q_SKIP,
   Q_CHMOD_FILE, Q_APPEND_FILE, Q_CAT_FILE, Q_DIFF_FILES,
   Q_SEND_QUIT, Q_CHANGE_USER, Q_MKDIR, Q_RMDIR,
+  Q_FORCE_RMDIR,
   Q_LIST_FILES, Q_LIST_FILES_WRITE_FILE, Q_LIST_FILES_APPEND_FILE,
   Q_SEND_SHUTDOWN, Q_SHUTDOWN_SERVER,
   Q_RESULT_FORMAT_VERSION,
@@ -464,6 +464,7 @@ const char *command_names[]=
   "disable_metadata",
   "exec",
   "execw",
+  "exec_in_background",
   "delimiter",
   "disable_abort_on_error",
   "enable_abort_on_error",
@@ -503,6 +504,7 @@ const char *command_names[]=
   "change_user",
   "mkdir",
   "rmdir",
+  "force-rmdir",
   "list_files",
   "list_files_write_file",
   "list_files_append_file",
@@ -3253,7 +3255,7 @@ static int replace(DYNAMIC_STRING *ds_str,
   mysqltest commmand(s) like "remove_file" for that
 */
 
-void do_exec(struct st_command *command)
+void do_exec(struct st_command *command, bool run_in_background)
 {
   int error;
   char buf[512];
@@ -3289,6 +3291,23 @@ void do_exec(struct st_command *command)
   while(replace(&ds_cmd, ">&-", 3, ">&4", 3) == 0)
     ;
 #endif
+  if (run_in_background)
+  {
+    /* Add an invocation of "START /B" on Windows, append " &" on Linux*/
+    DYNAMIC_STRING ds_tmp;
+#ifdef WIN32
+    init_dynamic_string(&ds_tmp, "START /B ",
+                        ds_cmd.length + 9, 256);
+   dynstr_append_mem(&ds_tmp, ds_cmd.str, ds_cmd.length);
+#else
+    init_dynamic_string(&ds_tmp, ds_cmd.str,
+                       ds_cmd.length + 2, 256);
+    dynstr_append_mem(&ds_tmp, " &", 2);
+#endif
+    dynstr_set(&ds_cmd, ds_tmp.str);
+    dynstr_free(&ds_tmp);
+  }
+
 
   /* exec command is interpreted externally and will not take newlines */
   while(replace(&ds_cmd, "\n", 1, " ", 1) == 0)
@@ -3302,17 +3321,19 @@ void do_exec(struct st_command *command)
     dynstr_free(&ds_cmd);
     die("popen(\"%s\", \"r\") failed", command->first_argument);
   }
-
-  while (fgets(buf, sizeof(buf), res_file))
+  if(!run_in_background)
   {
-    if (disable_result_log)
+    while (fgets(buf, sizeof(buf), res_file))
     {
-      buf[strlen(buf)-1]=0;
-      DBUG_PRINT("exec_result",("%s", buf));
-    }
-    else
-    {
-      replace_dynstr_append(&ds_res, buf);
+      if (disable_result_log)
+      {
+        buf[strlen(buf)-1]=0;
+        DBUG_PRINT("exec_result",("%s", buf));
+      }
+      else
+      {
+        replace_dynstr_append(&ds_res, buf);
+      }
     }
   }
   error= pclose(res_file);
@@ -3870,17 +3891,78 @@ void do_mkdir(struct st_command *command)
   DBUG_VOID_RETURN;
 }
 
+
+/*
+  SYNOPSIS
+  do_force_rmdir
+  command    - command handle
+  ds_dirname - pointer to dynamic string containing directory informtion
+
+  DESCRIPTION
+  force-rmdir <dir_name>
+  Remove the directory <dir_name>
+*/
+
+static void do_force_rmdir(struct st_command *command, DYNAMIC_STRING *ds_dirname)
+{
+  DBUG_ENTER("do_force_rmdir");
+
+  char dir_name[FN_REFLEN];
+  strncpy(dir_name, ds_dirname->str, sizeof(dir_name));
+
+  /* Note that my_dir sorts the list if not given any flags */
+  MY_DIR *dir_info= my_dir(ds_dirname->str, MYF(MY_DONT_SORT | MY_WANT_STAT));
+
+  if (dir_info && dir_info->number_off_files > 2)
+  {
+    /* Storing the length of the path to the file, so it can be reused */
+    size_t length= ds_dirname->length;
+
+    /* Delete the directory recursively */
+    for (uint i= 0; i < dir_info->number_off_files; i++)
+    {
+      FILEINFO *file= dir_info->dir_entry + i;
+
+      /* Skip the names "." and ".." */
+      if (!strcmp(file->name, ".") ||
+          !strcmp(file->name, ".."))
+        continue;
+
+      ds_dirname->length= length;
+      char dir_separator[2]= {FN_LIBCHAR, 0};
+      dynstr_append(ds_dirname, dir_separator);
+      dynstr_append(ds_dirname, file->name);
+
+      if (MY_S_ISDIR(file->mystat->st_mode))
+        /* It's a directory */
+        do_force_rmdir(command, ds_dirname);
+      else
+        /* It's a file */
+        my_delete(ds_dirname->str, MYF(0));
+    }
+  }
+
+  my_dirend(dir_info);
+  int error= rmdir(dir_name) != 0;
+  handle_command_error(command, error);
+
+  DBUG_VOID_RETURN;
+}
+
+
 /*
   SYNOPSIS
   do_rmdir
   command	called command
+  force         Recursively delete a directory if the value is set to true,
+                otherwise delete an empty direcory
 
   DESCRIPTION
   rmdir <dir_name>
   Remove the empty directory <dir_name>
 */
 
-void do_rmdir(struct st_command *command)
+static void do_rmdir(struct st_command *command, bool force)
 {
   int error;
   static DYNAMIC_STRING ds_dirname;
@@ -3894,8 +3976,13 @@ void do_rmdir(struct st_command *command)
                      ' ');
 
   DBUG_PRINT("info", ("removing directory: %s", ds_dirname.str));
-  error= rmdir(ds_dirname.str) != 0;
-  handle_command_error(command, error);
+  if (force)
+    do_force_rmdir(command, &ds_dirname);
+  else
+  {
+    error= rmdir(ds_dirname.str) != 0;
+    handle_command_error(command, error);
+  }
   dynstr_free(&ds_dirname);
   DBUG_VOID_RETURN;
 }
@@ -5223,7 +5310,7 @@ static void abort_process(int pid, const char *path)
 
 void do_shutdown_server(struct st_command *command)
 {
-  long timeout=90;
+  long timeout=600;
   int pid, error= 0;
   std::string ds_file_name;
   MYSQL* mysql = &cur_con->mysql;
@@ -5784,6 +5871,10 @@ void safe_connect(MYSQL* mysql, const char *name, const char *host,
                  "program_name", "mysqltest");
   mysql_options(mysql, MYSQL_OPT_CAN_HANDLE_EXPIRED_PASSWORDS,
                 &can_handle_expired_passwords);
+
+  set_server_public_key(mysql);
+  set_get_server_public_key_option(mysql);
+
   while(!mysql_real_connect(mysql, host,user, pass, db, port, sock,
                             CLIENT_MULTI_STATEMENTS | CLIENT_REMEMBER_OPTIONS))
   {
@@ -5889,6 +5980,10 @@ int connect_n_handle_errors(struct st_command *command,
   mysql_options4(con, MYSQL_OPT_CONNECT_ATTR_ADD, "program_name", "mysqltest");
   mysql_options(con, MYSQL_OPT_CAN_HANDLE_EXPIRED_PASSWORDS,
                 &can_handle_expired_passwords);
+
+  set_server_public_key(con);
+  set_get_server_public_key_option(con);
+
   while (!mysql_real_connect(con, host, user, pass, db, port, sock ? sock: 0,
                           CLIENT_MULTI_STATEMENTS))
   {
@@ -7142,6 +7237,7 @@ static struct my_option my_long_options[] =
    &sp_protocol, &sp_protocol, 0,
    GET_BOOL, NO_ARG, 0, 0, 0, 0, 0, 0},
 #include "sslopt-longopts.h"
+#include <caching_sha2_passwordopt-longopts.h>
   {"tail-lines", OPT_TAIL_LINES,
    "Number of lines of the result to include in a failure report.",
    &opt_tail_lines, &opt_tail_lines, 0,
@@ -8304,8 +8400,8 @@ void run_query_stmt(MYSQL *mysql, struct st_command *command,
   /* Init dynamic strings for warnings */
   if (!disable_warnings)
   {
-    init_dynamic_string(&ds_prepare_warnings, NULL, 0, 256);
-    init_dynamic_string(&ds_execute_warnings, NULL, 0, 256);
+    init_dynamic_string(&ds_prepare_warnings, "", 0, 256);
+    init_dynamic_string(&ds_execute_warnings, "", 0, 256);
   }
 
   /*
@@ -8407,13 +8503,6 @@ void run_query_stmt(MYSQL *mysql, struct st_command *command,
       append_stmt_result(ds, stmt, fields, num_fields);
 
       mysql_free_result(res);     /* Free normal result set with meta data */
-
-      /*
-        Clear prepare warnings if there are execute warnings,
-        since they are probably duplicated.
-      */
-      if (ds_execute_warnings.length || mysql->warning_count)
-        dynstr_set(&ds_prepare_warnings, NULL);
     }
     else
     {
@@ -8448,8 +8537,26 @@ void run_query_stmt(MYSQL *mysql, struct st_command *command,
           dynstr_append_mem(ds, ds_warnings->str,
                             ds_warnings->length);
         if (ds_prepare_warnings.length)
-          dynstr_append_mem(ds, ds_prepare_warnings.str,
-                            ds_prepare_warnings.length);
+        {
+          /* Split the string to get each warning */
+          std::stringstream prepare_warnings(ds_prepare_warnings.str);
+          std::string prepare_warning;
+          /*
+            If the warning is already present in the execute phase,
+            do not append it
+          */
+          while (std::getline(prepare_warnings, prepare_warning))
+          {
+            std::string execute_warnings(ds_execute_warnings.str);
+            if ((execute_warnings + "\n").find(prepare_warning + "\n") ==
+                std::string::npos)
+            {
+              dynstr_append_mem(ds, prepare_warning.c_str(),
+                                prepare_warning.length());
+              dynstr_append_mem(ds, "\n", 1);
+            }
+          }
+        }
         if (ds_execute_warnings.length)
           dynstr_append_mem(ds, ds_execute_warnings.str,
                             ds_execute_warnings.length);
@@ -9506,7 +9613,8 @@ int main(int argc, char **argv)
       case Q_REMOVE_FILE: do_remove_file(command); break;
       case Q_REMOVE_FILES_WILDCARD: do_remove_files_wildcard(command); break;
       case Q_MKDIR: do_mkdir(command); break;
-      case Q_RMDIR: do_rmdir(command); break;
+      case Q_RMDIR: do_rmdir(command, 0); break;
+      case Q_FORCE_RMDIR: do_rmdir(command, 1); break;
       case Q_LIST_FILES: do_list_files(command); break;
       case Q_LIST_FILES_WRITE_FILE:
         do_list_files_write_file_command(command, FALSE);
@@ -9718,7 +9826,11 @@ int main(int argc, char **argv)
         break;
       case Q_EXEC:
       case Q_EXECW:
-	do_exec(command);
+	do_exec(command, false);
+	command_executed++;
+	break;
+      case Q_EXEC_BACKGROUND:
+	do_exec(command, true);
 	command_executed++;
 	break;
       case Q_START_TIMER:
